add support for get requests to hyperclient
