+++ modules/hyperclient/src/testing.rs
@@ -28,7 +28,7 @@ use ethers::{
 
 use crate::{
 	internals,
-	internals::{request_status_stream, timeout_request_stream},
+	internals::{post_request_status_stream, timeout_post_request_stream},
 	providers::interface::Client,
 	types::{
 		ChainConfig, ClientConfig, EvmConfig, HashAlgorithm, MessageStatusWithMetadata,
@@ -135,7 +135,7 @@ pub async fn subscribe_to_request_status() -> Result<(), anyhow::Error> {
 	let block = receipt.block_number.unwrap();
 	tracing::info!("\n\nTx block: {block}\n\n");
 
-	let mut stream = request_status_stream(&hyperclient, post, block.low_u64()).await?;
+	let mut stream = post_request_status_stream(&hyperclient, post, block.low_u64()).await?;
 
 	while let Some(res) = stream.next().await {
 		match res {
@@ -258,11 +258,12 @@ pub async fn test_timeout_request() -> Result<(), anyhow::Error> {
 	let block = receipt.block_number.unwrap();
 	tracing::info!("\n\nTx block: {block}\n\n");
 
-	let request_status = request_status_stream(&hyperclient, post.clone(), block.low_u64()).await?;
+	let request_status =
+		post_request_status_stream(&hyperclient, post.clone(), block.low_u64()).await?;
 
 	// Obtaining the request stream and the timeout stream
 	let timed_out =
-		internals::request_timeout_stream(post.timeout_timestamp, hyperclient.source.clone()).await;
+		internals::message_timeout_stream(post.timeout_timestamp, hyperclient.source.clone()).await;
 
 	let mut stream = futures::stream::select(request_status, timed_out);
 
@@ -281,7 +282,7 @@ pub async fn test_timeout_request() -> Result<(), anyhow::Error> {
 		}
 	}
 
-	let mut stream = timeout_request_stream(&hyperclient, post).await?;
+	let mut stream = timeout_post_request_stream(&hyperclient, post).await?;
 
 	while let Some(res) = stream.next().await {
 		match res {
