+++ modules/ismp/pallets/state-coprocessor/src/lib.rs
@@ -77,7 +77,7 @@ pub mod pallet {
 			ensure_none(origin)?;
 
 			Self::handle_get_requests(message).map_err(|err| {
-				log::error!(target: "ismp", "{:?}", err);
+				log::error!(target: "ismp", "pallet-coprocessor: {:?}", err);
 				Error::<T>::HandlingError
 			})?;
